server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.7.0
google-genai==1.1.0
pyzbar==0.1.9
Pillow==11.1.0
httpx==0.28.1
aiosmtplib==3.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.11.0
python-multipart==0.0.20
pydantic==2.10.5
pydantic-settings==2.7.1
